pyproject.toml
# ============================================================================
# FILE: pyproject.toml
# ============================================================================
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "media-archive-toolkit"
version = "0.1.0"
description = "Migration and organization toolkit for media archives"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Rod Rivera", email = "rod@aiproduct.engineer" }
]

dependencies = [
    "blake3>=0.4.1",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "click>=8.1.7",
    "pathlib>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "black>=23.0.0",
]

[project.urls]
"Homepage" = "https://github.com/rodriveraai/media-vault"
"Bug Tracker" = "https://github.com/rodriveraai/media-vault/issues"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]